live_box_spreads/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "live-box-spreads"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dash>=2.15",
    "dash-bootstrap-components>=1.5",
    "plotly>=5.20",
    "polars>=0.20",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "requests>=2.31",
    "pyarrow>=15.0",
    "alpaca-py>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[project.scripts]
boxmon = "live_box_spreads.cli:main"
boxmon-ingest = "live_box_spreads.cli:main_ingest"
boxmon-dash = "live_box_spreads.cli:main_dashboard"

[tool.hatch.build.targets.wheel]
packages = ["src/live_box_spreads"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
